requirements.txt
pytz==2024.1
Django==4.2.11
djangorestframework==3.15.1
requests==2.31.0